requirements.txt
Django>=5.2
django-treebeard
